Interpolate glucose value on an equally spaced grid from day to day

Interpolate glucose value on an equally spaced grid from day to day

CGMS2DayByDay(data, dt0 = NULL, inter_gap = 45, tz = "")

Arguments

data
DataFrame object with column names "id", "time", and "gl". Should only be data for 1 subject. In case multiple subject ids are detected, the warning is produced and only 1st subject is used.

dt0
The time frequency for interpolation in minutes, the default will match the CGM meter's frequency (e.g. 5 min for Dexcom).

inter_gap
The maximum allowable gap (in minutes) for interpolation. The values will not be interpolated between the glucose measurements that are more than inter_gap minutes apart. The default value is 45 min.

tz
A character string specifying the time zone to be used. System-specific (see as.POSIXct), but " " is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning.

Value

A list with

gd2d

A matrix of glucose values with each row corresponding to a new day, and each column corresponding to time

actual_dates

A vector of dates corresponding to the rows of gd2d

dt0

Time frequency of the resulting grid, in minutes

Examples

CGMS2DayByDay(example_data_1_subject)

#> $gd2d
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,]        NA        NA        NA        NA        NA        NA        NA
#>  [2,] 112.17940 111.08696 110.08667 106.34667 106.00000 105.08667 102.26000
#>  [3,]  92.00000  92.92667  93.92667  94.92667  95.92667  96.00000  96.92667
#>  [4,]  90.06000  89.06000  90.88000  93.82000  94.00000  94.00000  94.00000
#>  [5,] 143.05298 138.23490 126.56000 117.42000 108.42000 102.28000  98.18667
#>  [6,]  76.06667  76.00000  75.03333  75.96667  77.93333  77.03333  77.96667
#>  [7,] 151.12000 144.14000 136.16000 129.14000 126.06000 127.96000 132.90000
#>  [8,]  96.99333  96.00667  95.00667  94.00667  93.00667  93.00000  92.00667
#>  [9,]  88.00667  89.00000  89.00000  89.00000  89.00000  89.00000  89.00000
#> [10,] 104.96000 102.94000 100.00000 100.02000 101.08000 105.02000 106.02000
#> [11,]  94.00000  94.03333  95.03333  96.03333  97.03333  98.00000  98.06667
#> [12,] 118.18667 122.04667 123.14000 126.09333 128.23333 133.04667 134.14000
#> [13,]  84.00000  83.94000  83.06000  84.12000  86.12000  88.18000  91.18000
#> [14,] 133.22000 136.44000 142.44000 148.51333 155.14667 156.85333 154.70667
#>            [,8]      [,9]     [,10]     [,11]     [,12]     [,13]     [,14]
#>  [1,]        NA        NA        NA        NA        NA        NA        NA
#>  [2,]  99.25084  96.25914  94.16722  93.08333  91.16667  89.17276  88.08667
#>  [3,]  97.92667  98.00000  96.14667  96.00000  94.14000  93.07000  92.07000
#>  [4,]  94.00000  94.00000  92.12000  91.06000  92.88000  92.05686  90.11960
#>  [5,]  95.14000  92.14000  90.09333  90.00000  90.00000  89.52333  89.02333
#>  [6,]  79.93333  80.96667  85.83333  99.53333 108.70000 116.73333 120.86667
#>  [7,] 136.92000 137.98000 137.02000 134.06000 134.00000 132.04000 130.04000
#>  [8,]  92.99333  93.00000  93.99333  94.00000  94.99333  94.00667  93.00667
#>  [9,]  89.00000  88.98667  87.00000  87.00000  87.01333  89.00667  90.00667
#> [10,] 107.00000 106.98000 106.02000 107.06000 110.04000 112.02000 113.06000
#> [11,] 100.00000 100.03333 101.03333 102.03333 103.00000 103.03333 104.03333
#> [12,] 137.18667 141.23333 146.04667 147.09333 149.04667 150.09333 152.14000
#> [13,]  94.24000  98.06000  99.06000  99.82000  96.82000  93.88000  92.00000
#> [14,] 150.85333 148.92667 147.92667 147.07333 147.41333 139.92667 138.48667
#>           [,15]     [,16]     [,17]     [,18]     [,19]     [,20]     [,21]
#>  [1,]        NA        NA        NA        NA        NA        NA        NA
#>  [2,]  88.00000  88.00000  87.08361  86.08333  87.83333  89.83333  91.83333
#>  [3,]  90.14618  90.00000  88.14000  88.93000  90.86000  91.93000  92.00000
#>  [4,]  91.88629  93.88667  94.00000  94.00000  94.94333  95.00000  94.05667
#>  [5,]  88.52333  88.02333  88.00000  88.95333  88.04348  87.04651  86.04348
#>  [6,] 126.80000 130.86667 133.90000 134.00000 132.06667 129.58333 127.08333
#>  [7,] 128.04000 122.12000 116.12000 110.10033 105.08333 103.03987 104.96656
#>  [8,]  93.99333  94.00000  94.00000  92.01333  92.00000  92.99333  94.98667
#>  [9,]  91.00667  92.00667  93.00667  94.00000  94.00667  95.00667  96.01338
#> [10,] 116.06000 119.06000 121.94000 118.94000 115.94000 112.94000 109.92000
#> [11,] 105.03333 106.03333 107.06667 109.00000 108.93333 107.00000 106.96667
#> [12,] 155.00000 155.04667 156.00000 156.00000 156.04667 157.00000 157.04667
#> [13,]  92.00000  92.00000  91.82000  88.94000  88.00000  88.18000  91.12000
#> [14,] 132.00000 132.44000 138.44000 144.44000 150.36667 155.07333 156.00000
#>           [,22]     [,23]     [,24]     [,25]     [,26]     [,27]     [,28]
#>  [1,]        NA        NA        NA        NA        NA        NA        NA
#>  [2,]  90.16667  89.08333  89.00000  88.08333  88.91667  89.91667  91.83333
#>  [3,]  92.92977  94.86000  95.00000  95.93000  96.00000  96.00000  95.07000
#>  [4,]  93.05667  92.05667  91.05667  90.05667  90.00000  90.00000  90.00000
#>  [5,]  86.00000  85.04333  84.04651  83.04667  83.95652  84.95349  85.00000
#>  [6,] 126.03010 125.03000 125.00000 127.91000 130.91000 131.97000 133.94000
#>  [7,] 105.98333 103.05000 101.03333 101.00000  99.03333  99.00000  97.03333
#>  [8,]  95.00000  96.99333  97.99667  98.99667  99.99667 100.00000 100.00000
#>  [9,]  98.01000  99.00000  99.00669 100.01000 101.00000 101.01000 102.01000
#> [10,] 105.92000 101.93980  98.97667  98.00000  98.00000  97.97667  96.95333
#> [11,] 106.00000 106.00000 105.93333 104.03344 104.96333 104.00000 104.00000
#> [12,] 158.04667 159.00000 159.00000 159.04667 160.09365 161.95000 161.00000
#> [13,]  93.00000  93.00000  92.88000  91.06000  92.06000  93.12000  95.12040
#> [14,] 156.36667 161.29333 165.51333 172.51333 179.22000 182.22000 185.22000
#>           [,29]     [,30]     [,31]     [,32]     [,33]     [,34]     [,35]
#>  [1,]        NA        NA        NA        NA        NA        NA        NA
#>  [2,]  92.00000  92.00000  93.83333  97.66667  98.00000  98.91667  99.91667
#>  [3,]  95.93000  96.93000  96.07000  95.07000  94.07000  94.00000  94.00000
#>  [4,]  90.00000  90.00000  90.94314  91.94333  92.94333  92.05667  92.00000
#>  [5,]  86.90667  89.86957  89.04333  89.00000  89.00000  89.95349  90.00000
#>  [6,] 135.93355 139.87960 143.88000 147.88000 149.94000 150.97000 152.94000
#>  [7,]  97.00000  96.01667  96.98333  96.01667  96.98333 100.93333 104.93333
#>  [8,]  97.01993  98.99331  99.99667 101.99333 101.00333 100.00333 100.00000
#>  [9,] 103.02000 105.02000 107.01000 108.01000 109.01000 110.01000 111.00000
#> [10,]  94.93023  91.93980  89.00000  88.97667  87.95333  86.00000  86.00000
#> [11,] 104.07333 106.07333 108.03667 109.00000 109.03667 109.89000 106.85333
#> [12,] 161.05000 162.04983 163.04682 164.00000 164.05000 165.00000 165.00000
#> [13,]  97.19000 100.12625 102.12040 104.12667 106.12667 108.12667 110.06333
#> [14,] 188.22074 191.15333 193.23000 196.53667 203.53667 210.30667 214.00000
#>           [,36]     [,37]     [,38]    [,39]     [,40]     [,41]     [,42]
#>  [1,]        NA        NA        NA       NA        NA        NA        NA
#>  [2,] 100.91667 101.91667 102.91667 103.9167 104.91667 105.91667 105.08638
#>  [3,]  94.00000  94.00000  94.00000  94.0000  94.00000  94.00000  94.00000
#>  [4,]  92.00000  91.05667  91.00000  91.0000  91.00000  90.05686  90.94333
#>  [5,]  90.95667  90.04651  90.00000  90.0000  90.00000  91.91304  92.95667
#>  [6,] 151.06000 145.18000 140.15000 138.0600 134.12000 128.18000 122.18000
#>  [7,] 105.00000 105.98333 103.05000 103.9833 104.98333 106.96667 108.96667
#>  [8,]  98.00667  98.00000  98.99336  99.0000  96.01000  96.00000  97.99333
#>  [9,] 111.01000 112.01000 113.01000 114.0000 114.00000 114.00000 114.01000
#> [10,]  85.97667  84.97667  83.95333  82.0000  81.97667  81.02333  82.04667
#> [11,] 102.89000  99.92667  97.96333  97.0000  97.03667  98.03667  99.03667
#> [12,] 164.90000 162.90000 161.00000 160.8500 157.80000 153.75000 148.95000
#> [13,] 111.06333 112.06333 112.81000 109.7467 105.81063 102.93980 101.93667
#> [14,] 214.00000 213.84667 211.54000 205.6933 201.84667 199.69333 195.92333
#>           [,43]     [,44]     [,45]     [,46]     [,47]     [,48]     [,49]
#>  [1,]        NA        NA        NA        NA        NA        NA        NA
#>  [2,] 105.91639 104.16667 102.16667 100.16667 100.00000 101.83333 102.00000
#>  [3,]  94.00000  93.07000  92.07000  93.86000  94.93000  95.46500  95.96500
#>  [4,]  91.00000  91.94333  92.00000  92.94333  93.94333  95.88040  96.00000
#>  [5,]  93.00000  94.90698  95.95652  94.08667  94.00000  94.95349  95.95652
#>  [6,] 119.09000 116.09000 113.09000 111.06000 108.09967 106.06020 105.03000
#>  [7,] 111.95000 113.96667 116.95000 119.95000 120.98333 121.98333 123.96667
#>  [8,] 100.99000 100.00333  98.00667  97.00333  97.00000  96.00333  95.00333
#>  [9,] 115.00000 115.00000 115.01000 116.01000 117.00000 117.00000 116.98000
#> [10,]  84.02333  85.00000  85.00000  85.07000  88.02333  89.00000  89.02333
#> [11,] 100.11000 103.07333 105.11000 108.03667 109.03667 110.00000 110.00000
#> [12,] 148.00000 148.00000 148.00000 147.95000 146.95000 146.00000 145.95000
#> [13,] 101.00000 101.06020 102.06333 102.81000 100.00000 100.06333 101.12667
#> [14,] 194.92333 193.92333 192.92333 192.15333 194.15333 195.92333 194.77000
#>           [,50]     [,51]     [,52]     [,53]     [,54]     [,55]     [,56]
#>  [1,]        NA        NA        NA        NA        NA        NA        NA
#>  [2,] 102.91667 103.91667 104.91667 104.08333 100.33333  97.25000  96.08333
#>  [3,]  96.00000  96.00000  96.00000  96.00000  96.00000  96.93000  96.07000
#>  [4,]  96.00000  95.05667  95.00000  94.05667  94.00000  94.00000  94.00000
#>  [5,]  96.00000  96.95667  97.00000  96.04333  95.04651  94.04348  92.08667
#>  [6,] 103.06000 101.06000  92.27000  87.15000  84.09000  86.91000  93.79000
#>  [7,] 125.96667 126.98333 127.98333 128.00000 127.01667 127.00000 128.96667
#>  [8,]  95.00000  95.00000  94.00333  94.00000  94.00000  94.99667  95.00000
#>  [9,] 115.00000 115.00000 114.98000 113.01000 114.01000 115.01000 116.00000
#> [10,]  89.97667  89.00000  89.02333  90.02333  91.02333  92.00000  92.02333
#> [11,] 109.96333 108.96333 108.00000 107.96333 107.00000 107.00000 107.03667
#> [12,] 145.05000 146.15000 149.15000 152.10000 154.00000 154.10000 156.05000
#> [13,] 103.06333 104.06333 105.00000 105.06333 106.00000 106.06333 107.06333
#> [14,] 191.54000 185.31000 176.46333 169.15667 159.07667 159.61667 154.69333
#>           [,57]     [,58]     [,59]     [,60]     [,61]     [,62]     [,63]
#>  [1,]        NA        NA        NA        NA        NA        NA        NA
#>  [2,]  96.00000 102.41667 106.66667 107.00000 106.08333 104.16667 103.08333
#>  [3,]  95.07000  95.00000  95.93000  96.93000  98.86000  99.00000  99.00000
#>  [4,]  93.05667  93.00000  93.00000  93.00000  93.00000  93.94333  95.88667
#>  [5,]  93.91333  95.91333  97.91333  98.95667  99.95667 100.95667 101.00000
#>  [6,]  94.00000  94.00000  94.97000  95.00000  95.97000  97.94000  98.97000
#>  [7,] 130.96667 130.01667 130.98333 131.98333 132.98333 135.95000 138.95000
#>  [8,]  95.00000  95.99667  96.99667  97.00000  97.99667  98.00000  98.99667
#>  [9,] 116.01000 117.01000 118.02000 120.01000 121.01000 122.00000 122.01000
#> [10,]  93.00000  92.97667  92.02333  93.00000  93.02333  94.00000  94.04667
#> [11,] 108.07333 110.11000 113.11000 116.07333 118.03667 119.00000 118.89000
#> [12,] 156.90000 154.90000 152.90000 150.90000 148.90000 147.15000 149.95000
#> [13,] 108.00000 108.12667 110.12667 112.19000 115.12667 117.19000 120.12667
#> [14,] 150.77000 147.77000 145.07667 146.07667 147.07667 147.84667 145.69333
#>           [,64]     [,65]     [,66]    [,67]    [,68]     [,69]    [,70]
#>  [1,]        NA        NA        NA       NA       NA        NA       NA
#>  [2,] 103.91667 104.00000 104.00000 104.0000 104.0000 104.00000 104.0000
#>  [3,]  98.07000  97.07000  97.93000  98.0000  97.0700  97.00000  94.2100
#>  [4,]  96.94333  98.88667  99.94333 100.0000 100.0000  99.05667  99.0000
#>  [5,] 101.00000 100.04333 100.95667 101.9567 102.0000 101.04333 101.0000
#>  [6,]  99.97000 101.94000 103.94000 103.0300 102.0300 104.91000 106.9400
#>  [7,] 141.95000 144.95000 146.96667 148.9667 150.9667 150.01667 147.0500
#>  [8,]  99.00000  99.00000 100.99333 101.9967 103.9933 104.99667 105.9967
#>  [9,] 123.01000 123.98000 122.00000 122.0000 122.0100 123.02000 125.0100
#> [10,]  96.02333  97.04667  99.04667 101.0000 101.0233 102.04667 104.0467
#> [11,] 115.96333 114.96333 113.96333 112.9633 111.9633 110.96333 110.0000
#> [12,] 149.10000 151.00000 151.00000 151.0000 151.0000 151.05000 151.9500
#> [13,] 122.00000 122.06333 123.00000 122.8733 120.8733 118.87333 116.9367
#> [14,] 141.84667 139.92333 138.92333 138.0000 138.0000 137.38667 129.6933
#>          [,71]    [,72]    [,73]     [,74]    [,75]    [,76]     [,77]    [,78]
#>  [1,]       NA       NA       NA        NA       NA       NA        NA       NA
#>  [2,] 103.0833 102.0833 100.1667 101.83333 102.9167 104.8333 105.91667 106.9167
#>  [3,]  94.0000  95.8600  96.0000  94.14000  97.7200  98.9300  99.93000 100.9300
#>  [4,]  99.0000  99.0000  99.0000  99.94333 100.9433 100.0567  99.05667  99.0000
#>  [5,] 101.9567 102.0000 102.9567 103.00000 103.0000 102.5217 102.02167 102.0000
#>  [6,] 108.9400 109.9700 110.9700 111.97000 112.0000 113.9398 116.91000 117.9700
#>  [7,] 144.0500 142.0333 142.9833 143.98333 144.0000 144.9833 145.00000 145.9833
#>  [8,] 106.0000 105.0033 105.0000 105.00000 105.0000 105.0000 105.99667 106.0000
#>  [9,] 126.0000 126.0000 126.0100 127.00000 126.9900 126.0100 126.99000 125.9900
#> [10,] 105.9767 105.0233 106.0233 107.02333 108.0000 107.9767 106.97667 106.0000
#> [11,] 110.0367 111.0367 112.0367 113.00000 112.9633 112.0000 112.00000 112.0000
#> [12,] 151.0500 152.0000 152.0500 153.00000 153.0000 152.9000 150.95000 149.9500
#> [13,] 115.8733 113.9367 113.0633 114.00000 113.8733 112.0000 111.93667 110.8733
#> [14,] 126.7667 135.8467 134.0000 133.92333 132.8467 131.0767 132.07667 133.0000
#>          [,79]    [,80]     [,81]    [,82]    [,83]    [,84]    [,85]    [,86]
#>  [1,]       NA       NA        NA       NA       NA       NA       NA       NA
#>  [2,] 107.0000 107.0000 107.91667 109.8395 110.9200 111.9200 112.9200 113.0000
#>  [3,] 101.0000 101.0000 103.79000 104.9269 105.9298 106.9331 107.0000 107.9333
#>  [4,] 100.8867 101.9433 102.00000 102.9433 102.0567 102.0000 102.0000 102.9465
#>  [5,] 101.0433 100.0433  99.04333 100.9133 103.8700 104.0000 105.9133 106.9567
#>  [6,] 118.9700 119.0000 119.00000 120.9400 121.9700 122.9700 123.0000 123.0000
#>  [7,] 147.9667 148.9833 149.98333 150.0000 150.0000 150.0000 148.0333 147.0167
#>  [8,] 104.0067 102.0067 101.00333 101.0000 100.0033 100.9967 102.9933 103.9967
#>  [9,] 125.0100 126.0100 127.00000 127.0100 128.0100 129.0000 128.9900 128.0000
#> [10,] 106.0233 107.0233 108.00000 108.0000 108.0233 109.0233 110.0000 110.0233
#> [11,] 111.9633 110.9267 108.96333 108.0000 107.9633 107.0000 107.0000 107.0000
#> [12,] 149.0500 150.0500 151.05000 152.1000 154.2000 157.9000 155.7500 150.7500
#> [13,] 108.8733 106.9367 105.93667 105.0633 105.9367 105.0633 105.9367 104.7467
#> [14,] 133.1533 135.0767 135.92333 134.8467 133.0000 133.2300 135.9233 135.0000
#>          [,87]     [,88]     [,89]    [,90]    [,91]    [,92]    [,93]    [,94]
#>  [1,]       NA        NA        NA       NA       NA       NA       NA       NA
#>  [2,] 113.0000 113.91973 114.92000 114.0800 115.8400 116.0000 115.0800 116.8400
#>  [3,] 108.9302 109.00000 109.93023 110.0000 110.0000 109.0667 108.0667 107.0667
#>  [4,] 103.0000 103.00000 103.00000 103.9467 104.9467 105.9467 106.0000 107.8933
#>  [5,] 107.0000 107.00000 108.92000 109.9568 110.9599 107.1728 105.0803 105.9536
#>  [6,] 123.0000 122.03000 121.03000 121.9700 122.9700 123.9700 124.0000 122.0533
#>  [7,] 146.0167 149.93333 152.93046 153.0000 153.9833 159.9197 171.8400 180.8800
#>  [8,] 102.0067 102.00000 102.00000 102.0000 102.9967 103.0000 103.0000 103.0000
#>  [9,] 128.0200 130.00000 129.99000 128.9900 127.9800 125.9900 124.9900 123.9900
#> [10,] 111.0000 110.95333 108.95333 107.0467 109.0467 110.9767 109.9533 107.9767
#> [11,] 106.9633 106.03667 106.96333 106.0000 106.0367 107.0000 106.9267 104.9633
#> [12,] 145.7500 140.75166 135.95638 135.0000 135.0500 136.0997 138.3278 145.2000
#> [13,] 100.7467  97.06333  98.50667 106.3167 111.1267 113.0633 114.1267 116.0633
#> [14,] 134.6933 130.84768 129.00000 129.0767 130.0767 131.1533 133.3067 136.6933
#>          [,95]    [,96]    [,97]    [,98]    [,99]   [,100]   [,101]   [,102]
#>  [1,]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [2,] 118.8400 118.0800 109.7200 102.5600  99.2400 102.6800  99.3200  95.3200
#>  [3,] 106.0667 105.0667 103.1333 104.8667 105.0000 108.7333 109.0000 104.3654
#>  [4,] 108.9467 108.0533 108.0000 106.1192 105.0569 104.0567 104.9465 105.0000
#>  [5,] 106.9597 108.9200 109.9600 111.9200 113.9200 116.8800 122.7600 125.8800
#>  [6,] 120.0533 113.1867 113.0000 113.0000 113.0000 113.0000 113.9732 115.9467
#>  [7,] 188.8933 197.8505 199.9732 203.9467 208.9333 209.9867 207.0400 205.0267
#>  [8,] 102.0033 100.0000 110.9635 112.0000 110.0000 109.0000 109.0000 109.0000
#>  [9,] 123.0000 122.9867 122.0000 121.9867 121.0000 120.9867 119.9900 118.9467
#> [10,] 106.9767 106.0000 106.0000 106.0000 106.0533 108.0506 110.0479 112.0453
#> [11,] 103.9633 103.0000 103.0367 104.0367 105.0611 106.7278 108.3944 110.1100
#> [12,] 149.1000 150.8000 146.8000 142.9000 140.9000 138.7993 134.9468 133.8495
#> [13,] 117.0633 118.0633 119.0633 120.0633 120.6833 115.8422 113.3505 111.0000
#> [14,] 132.7700 130.0000 130.0000 129.3100 120.7700 117.7700 114.9233 114.0000
#>         [,103]   [,104]   [,105]   [,106]   [,107]   [,108]   [,109]   [,110]
#>  [1,]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [2,]  94.0800  94.9200 100.5200 101.0000 101.0000  99.1600  99.0000  99.0000
#>  [3,] 104.0000 104.9333 105.9333 107.8667 109.8667 113.7333 116.7907 118.8600
#>  [4,] 106.8933 108.8933 111.8400 112.9467 113.9467 116.5244 119.1911 121.8578
#>  [5,]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [6,] 117.9467 123.8400 129.8400 133.8869 137.8802 138.9732 139.9733 141.9467
#>  [7,] 205.9867 204.0267 199.0667 196.0400 196.0000 194.0267 192.0267 187.0667
#>  [8,] 111.0000 112.0000 116.0000 120.0000 122.0000 123.0000 124.0000 125.5000
#>  [9,] 114.9933 114.4942 113.9699 111.0000 111.0000 111.0133 112.0400 115.0400
#> [10,] 114.0426 116.0201 117.0267 118.0267 119.0266 120.0000 120.0000 119.9467
#> [11,] 113.0513 114.4513 115.8513 117.2513 118.6513 119.9264 117.8007 113.0000
#> [12,] 130.6267 123.9467 122.7333 117.6811 111.8997 110.0533 111.0533 111.9467
#> [13,] 111.0000 111.1500 113.4000 115.6500 117.9000 120.1329 122.0317 122.5317
#> [14,] 114.0769 115.0000 115.0800 116.0000 115.9200       NA       NA       NA
#>         [,111]   [,112]   [,113]   [,114]   [,115]   [,116]   [,117]   [,118]
#>  [1,]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [2,]  99.0000  99.9200 101.8400 103.8400 106.7508 107.9197 108.0000 109.8364
#>  [3,] 120.8600 121.9300 122.9300 123.0000 123.0000 123.0000 123.0000 123.0000
#>  [4,] 122.9467 122.0565 121.0582 120.0599 120.0000 120.0000 119.0533 119.0000
#>  [5,]       NA       NA       NA       NA 115.5650 114.0650 113.0401 110.1200
#>  [6,] 142.9733 143.9733 141.0800 139.0533 137.0533 132.1333 129.0897 127.0535
#>  [7,] 181.0800 172.1200 163.1200 163.0000 163.9867 163.0133 164.9733 164.0133
#>  [8,] 127.0000 127.0000 126.0000 124.0000 122.0000 120.0000 119.0000 118.0000
#>  [9,] 118.0267 120.0133 121.0000 120.9933 120.4933 120.0000 120.0000 120.0000
#> [10,] 117.9467 115.9467 113.9467 112.0000 111.9733 110.9467 109.0000 109.0000
#> [11,] 112.9200 111.1200 114.0400 114.8400 110.9200 108.9200 107.0800 109.1600
#> [12,] 111.0000 110.8933 109.1067 111.0533 112.1600 115.2133 119.1600 122.1600
#> [13,] 123.0633 124.0633 125.0633 126.0633 126.7467 122.7458 118.8667 116.8667
#> [14,]       NA       NA       NA       NA       NA       NA       NA       NA
#>         [,119]   [,120]   [,121]   [,122]   [,123]   [,124]    [,125]    [,126]
#>  [1,]       NA       NA       NA       NA       NA       NA        NA        NA
#>  [2,] 111.8397 112.0000 112.0000 112.0000 111.0800 109.7733 108.44000 107.10667
#>  [3,] 123.0000 121.1338 118.2000 115.2000 112.2000 110.1333 110.00000 108.13333
#>  [4,] 118.6852 118.3515 118.0178 116.1067 116.0000 114.5800 113.08000 110.79356
#>  [5,] 109.0400 107.4000 105.7333 104.0667 102.4000 100.7333  99.06667  99.96000
#>  [6,] 124.0800 120.1067 117.0800 113.1067 109.1067 107.0533 105.05333 102.08000
#>  [7,] 159.0993 155.0537 154.3422 153.6756 153.0089 151.0267        NA        NA
#>  [8,] 116.0000 114.0000 114.0000 114.0000 112.0000 112.0000 112.00000 109.00000
#>  [9,] 119.9867 118.9867 117.9867 116.9867 116.0000 116.0000 115.98667 115.00000
#> [10,] 108.9467 107.0000 106.9734 105.9766 104.9734 103.9766 102.97333 101.97333
#> [11,] 112.8800 109.9200 107.8800 105.0400 106.0400 107.0000 107.12000 110.08000
#> [12,] 125.0533 125.7867 121.7333 116.7333 111.8933 110.0533 111.10667 113.21333
#> [13,] 114.9333 113.8667 111.9333 110.8000 107.8667 105.8000 102.80000  99.86667
#> [14,]       NA       NA       NA       NA       NA       NA        NA        NA
#>         [,127]    [,128]    [,129]    [,130]    [,131]    [,132]    [,133]
#>  [1,]       NA        NA        NA        NA        NA        NA        NA
#>  [2,] 106.3867 105.72000 105.05333 110.52000 109.62000 108.12000 107.08000
#>  [3,] 107.0667 106.06667 105.06667 103.13333 101.13333  98.20000  96.75556
#>  [4,] 108.4628 106.13208 106.00000 105.05667 104.52794 104.02752 103.52711
#>  [5,]  99.0400  98.04000  96.08000  94.08638  93.36151  92.69410  92.02670
#>  [6,] 105.8933 109.89333 113.89333 114.97333 112.08000 109.08000 109.00000
#>  [7,]       NA        NA        NA        NA        NA        NA        NA
#>  [8,] 108.0000 107.00000 107.00000 107.00000 106.00000 104.00000 104.00000
#>  [9,] 115.0000 115.00000 115.00000 115.00000 114.97333 112.97333 110.96000
#> [10,] 101.0000 101.00000 101.00000 100.94667  98.97333  98.00000  97.94667
#> [11,] 111.9200 110.03987 110.96321 109.88000 107.00000 107.00000 107.08000
#> [12,] 117.1067 119.21262 123.25084 128.26667 133.21333 137.26667 142.10667
#> [13,]  98.0000  97.93333  96.86711  94.81000  91.87291  89.93333  89.00000
#> [14,]       NA        NA        NA        NA        NA        NA        NA
#>          [,134]    [,135]    [,136]    [,137]    [,138]    [,139]    [,140]
#>  [1,]        NA        NA        NA        NA        NA        NA        NA
#>  [2,] 107.00000 106.08000 106.00000 105.54000 105.04000 105.46000 105.96000
#>  [3,]  95.42222  94.08889        NA        NA        NA        NA        NA
#>  [4,] 103.02669 102.36959 101.70366 101.03774 100.05351  99.05333  99.00000
#>  [5,]  92.48000  92.98000  91.08000  90.04000  89.52000  89.02000  88.04000
#>  [6,] 109.00000 109.97333 110.48667 110.98667 110.02667 109.02667 109.97333
#>  [7,]        NA        NA        NA        NA        NA        NA        NA
#>  [8,] 105.00000 103.00000 102.00000 101.00000 101.00000 100.00000 100.00000
#>  [9,] 107.96000 104.97333 103.02667 105.00000 104.97333 103.00000 102.98667
#> [10,]  96.00000  95.97342  94.95349  92.95973  90.94667  89.00000  89.08000
#> [11,] 109.12000 112.12000 115.08000 117.16000 121.20000 126.12000 129.12000
#> [12,] 144.16000 147.21333 151.10667 152.89333 150.89333 148.78667 144.84000
#> [13,]  89.06667  90.00000  89.93333  88.86667  86.93333  85.93333  84.86667
#> [14,]        NA        NA        NA        NA        NA        NA        NA
#>          [,141]    [,142]    [,143]    [,144]    [,145]    [,146]    [,147]
#>  [1,]        NA        NA        NA        NA        NA        NA        NA
#>  [2,] 105.08000 105.00000 105.91973 102.32000  99.24000  99.13137  99.27416
#>  [3,]        NA        NA        NA        NA        NA        NA        NA
#>  [4,]  96.16000  94.10667  92.10667  91.05333  91.94667  92.47333  92.97333
#>  [5,]  87.04000        NA        NA        NA        NA        NA        NA
#>  [6,] 112.92000 111.05333 107.10667 105.05333 102.08000 101.02667  99.05333
#>  [7,]        NA        NA        NA        NA        NA        NA        NA
#>  [8,]  97.00000  96.00000  95.00000  94.00000  95.00000  94.00000  93.00000
#>  [9,] 102.00667 102.50667 102.98667 101.99000 101.24000 100.49000  99.74000
#> [10,]  92.08000  95.05333  97.02667  98.05333 100.05333 102.05333 104.02667
#> [11,] 132.71761 149.92642 148.16000 152.04000 153.08000 155.04000 156.16000
#> [12,] 142.00000 142.05333 143.10667 145.16000 148.21333 152.16000 154.94667
#> [13,]  82.86667  81.06667  82.06667  83.13333  85.13333  87.06667  87.80000
#> [14,]        NA        NA        NA        NA        NA        NA        NA
#>          [,148]    [,149]    [,150]    [,151]    [,152]    [,153]    [,154]
#>  [1,]        NA        NA        NA        NA        NA        NA        NA
#>  [2,]  99.41694  99.55973  99.70252  99.84531  99.98810  99.07718  99.00000
#>  [3,]        NA        NA        NA        NA        NA        NA        NA
#>  [4,]  93.47333  93.97333  94.00000  94.00000  94.00000  94.00000  93.05457
#>  [5,]        NA        NA        NA        NA        NA        NA        NA
#>  [6,]  96.08000  95.02667  96.94667  93.10667  91.05333  90.02667  90.00000
#>  [7,]        NA        NA        NA        NA        NA        NA        NA
#>  [8,]  92.00000  92.00000  91.00000  92.00000  92.00000  94.00000  93.00000
#>  [9,]  98.99000  98.24000  97.49000  96.74000  95.97333  93.98667  93.02667
#> [10,] 105.05333 107.02667 108.02667 109.00000 108.97333 107.97333 106.94667
#> [11,] 160.04000 161.04000 162.08000 164.04000 165.04000 165.96000 164.92000
#> [12,] 154.10667 156.05333 157.10667 159.10667 160.89333 158.87556 156.54222
#> [13,]  84.93333  84.06667  84.86667  83.00000  83.00000  83.13333  85.20000
#> [14,]        NA        NA        NA        NA        NA        NA        NA
#>          [,155]    [,156]    [,157]    [,158]    [,159]    [,160]    [,161]
#>  [1,]        NA        NA        NA        NA        NA        NA        NA
#>  [2,]  97.16000  97.00000  97.00000  96.53833  96.03833  98.76080 100.53333
#>  [3,]        NA        NA        NA        NA        NA        NA        NA
#>  [4,]  92.05234  91.05011  91.00000  91.00000  91.00000        NA        NA
#>  [5,]        NA        NA        NA        NA        NA        NA        NA
#>  [6,]  88.05333  87.02667  85.05333  85.97333  86.97333  86.02667  86.00000
#>  [7,]        NA        NA        NA        NA        NA        NA        NA
#>  [8,]  94.00000  93.00000  93.00000  93.50000  94.00000  92.00000  92.00000
#>  [9,]  95.02667  96.96000  94.00000  94.01333  95.01333  96.01333  97.00000
#> [10,] 105.00000 104.94667 102.92000  99.92000  96.92000  93.94667  91.97333
#> [11,] 162.72000 155.84000 151.84000 147.80000 142.68000 134.68000 126.72000
#> [12,] 154.20889 151.62667 144.73333 140.05333 141.10667 143.10667 145.00000
#> [13,]  88.13333  90.06667  91.20000  94.33333  99.80000 111.73333 122.26667
#> [14,]        NA        NA        NA        NA        NA        NA        NA
#>          [,162]    [,163]    [,164]    [,165]   [,166]    [,167]    [,168]
#>  [1,]        NA        NA        NA        NA       NA        NA        NA
#>  [2,] 102.20000 103.86667 105.84615 106.92333 107.9233 108.92180 109.92013
#>  [3,]        NA        NA        NA        NA       NA 182.80936 179.25333
#>  [4,]        NA        NA        NA        NA       NA        NA        NA
#>  [5,]        NA        NA        NA        NA       NA        NA        NA
#>  [6,]  86.00000  86.97342  87.00000  87.00000  87.0000  86.02333  86.97342
#>  [7,]        NA        NA        NA        NA       NA        NA 120.98997
#>  [8,]  93.00000  93.00000  94.00000  95.00000  95.0000  94.00000  95.00000
#>  [9,]  96.98667  96.04000  99.06667 104.06667 109.1600 121.17391 134.06667
#> [10,]  90.94667  88.94667  86.97333  85.97333  85.0000  85.02676  85.94000
#> [11,] 119.84000 115.88000 112.92000 110.88000 107.9600 106.96000 105.96000
#> [12,] 145.00000 144.89333 143.05333 144.05333 145.0533 145.94667 145.16000
#> [13,] 126.53333 134.46667 141.53333 149.26667 153.2000 156.06667 157.06667
#> [14,]        NA        NA        NA        NA       NA        NA        NA
#>          [,169]    [,170]    [,171]    [,172]    [,173]    [,174]    [,175]
#>  [1,]        NA        NA        NA        NA        NA        NA        NA
#>  [2,] 110.92154 111.92321 112.61556 113.28222 113.94889 114.92333 115.00000
#>  [3,] 178.06645 175.20930 174.06845 173.06678 168.31773 157.73090 149.52835
#>  [4,]        NA        NA        NA        NA        NA        NA        NA
#>  [5,]        NA        NA        NA        NA        NA        NA        NA
#>  [6,]  87.00000  87.97667  88.97667  89.97667  90.00000  88.04667  88.97667
#>  [7,] 121.99000 123.98000        NA        NA        NA        NA        NA
#>  [8,]  95.00667  96.99336  95.00000  96.00000  96.50083  96.99668  96.00000
#>  [9,] 138.00000 137.95000 134.90864 129.42691 123.93960 117.95833 115.45833
#> [10,]  83.94000  82.03000  83.06000  85.33000  96.26910 105.13378 110.15000
#> [11,] 105.04000 106.04013 106.91333 104.87000 101.87043  98.87960  96.00000
#> [12,] 148.00000 148.00000 148.05351 149.00000 149.22667 153.11333 155.17000
#> [13,] 158.40000 164.40000 170.40000 176.46667 183.26667 187.66890 197.63000
#> [14,]        NA        NA        NA        NA        NA        NA        NA
#>          [,176]    [,177]    [,178]    [,179]    [,180]    [,181]    [,182]
#>  [1,]        NA        NA        NA        NA        NA        NA        NA
#>  [2,] 114.53833 114.03833        NA        NA        NA        NA        NA
#>  [3,] 141.52302 133.51768 125.51234 117.50700 109.50667 101.50667  93.50667
#>  [4,]        NA        NA        NA        NA        NA        NA        NA
#>  [5,]        NA        NA        NA        NA        NA        NA        NA
#>  [6,]  88.02658  88.00000  88.00000  87.02333  89.34400  91.74400  94.14400
#>  [7,]        NA        NA        NA        NA        NA        NA        NA
#>  [8,]  95.00000  95.00000  94.99668  94.00000  93.99667  92.99667  92.00333
#>  [9,] 113.00000 113.00000 113.03333 115.07500 119.57500 124.04167 126.54167
#> [10,] 115.12000 119.15000 124.09000 127.00000 126.94000 125.06000 127.15000
#> [11,]  96.08667  98.17333 102.21667 107.26000 113.13000 116.13000 119.08667
#> [12,] 158.25500 162.75500 167.16944 170.05351 171.11333 172.88667 171.00000
#> [13,] 206.28000 210.14000 211.93000 210.86000 208.86047 206.73244 202.79000
#> [14,]        NA        NA        NA        NA        NA        NA        NA
#>          [,183]    [,184]   [,185]   [,186]   [,187]    [,188]    [,189]
#>  [1,]        NA        NA       NA       NA       NA        NA        NA
#>  [2,]        NA        NA       NA       NA       NA        NA        NA
#>  [3,]  86.44333  81.31667  78.1900  76.5950  75.0950  73.12667  71.12978
#>  [4,]        NA 103.95000 102.1000  99.1500 101.8500 102.00000 103.89967
#>  [5,]        NA        NA       NA       NA       NA        NA        NA
#>  [6,]  96.54400  98.94400 115.6033 142.3700 153.7433 169.62667 193.44000
#>  [7,]        NA        NA       NA       NA       NA        NA        NA
#>  [8,]  93.01000  96.01333 100.0000 100.0133 104.0033 105.00667 106.99889
#>  [9,] 129.06667 133.05000 136.0166 136.9867 136.0134 137.00000 137.03333
#> [10,] 132.12000 136.00000 136.1200 140.0600 142.1800 148.18000 154.12000
#> [11,] 121.04333 121.95667 120.9133 118.9567 118.0000 118.04333 119.08667
#> [12,] 170.83000 167.77333 163.6600 157.8300 154.8867 153.17000 156.17000
#> [13,] 199.79000 196.86000 194.8600 193.0700 194.0700 194.93000 193.93000
#> [14,]        NA        NA       NA       NA       NA        NA        NA
#>          [,190]    [,191]    [,192]   [,193]    [,194]    [,195]    [,196]
#>  [1,]        NA        NA        NA       NA        NA        NA        NA
#>  [2,]        NA        NA        NA       NA        NA        NA        NA
#>  [3,]  69.13311  67.12709  66.06333  66.0000  68.18556  70.51889  72.85222
#>  [4,] 104.95000 111.65000 112.00000 112.0000 119.60000 125.70000 126.95000
#>  [5,]        NA        NA        NA       NA        NA        NA        NA
#>  [6,] 218.41667 236.58000 254.58000 270.6267 275.88333 273.07000 263.23333
#>  [7,]        NA        NA        NA       NA        NA        NA        NA
#>  [8,] 106.66556 106.33222 105.99889 105.6656 105.33222 105.00332 106.00000
#>  [9,] 139.03333 141.08333 146.08333 151.0667 155.11667 162.11667 169.09967
#> [10,] 158.12000 162.12000 166.18000 172.1200 176.12000 180.06000 182.09000
#> [11,] 121.04333 122.00000 121.95667 121.1733 125.08667 127.04333 128.04333
#> [12,] 159.05667 160.17000 163.17000 166.1700 169.17000 171.94352 170.67893
#> [13,] 193.14000 195.07000 196.14000 198.0000 198.13953 200.06667 201.13378
#> [14,]        NA        NA        NA       NA        NA        NA        NA
#>         [,197]  [,198]   [,199]   [,200]   [,201]   [,202]   [,203]   [,204]
#>  [1,]       NA      NA       NA       NA       NA       NA 148.1467 142.8133
#>  [2,]       NA      NA       NA       NA       NA       NA       NA       NA
#>  [3,]       NA      NA       NA       NA       NA       NA       NA       NA
#>  [4,] 126.0500 129.800 132.8500 135.8500 139.8000 140.9500 142.9000 143.9500
#>  [5,]       NA      NA       NA       NA       NA       NA       NA       NA
#>  [6,] 243.4667 241.535 240.0350 224.3733 199.5833 187.2800 173.3267 164.2100
#>  [7,]       NA      NA       NA       NA       NA       NA       NA       NA
#>  [8,] 114.0332 124.000 132.0333 142.0465 156.0000 172.0267 180.0017 180.5017
#>  [9,] 175.0800 181.080 187.0401 190.0083 190.5083 191.0000 190.9167 185.9167
#> [10,] 185.1200 189.120 193.0300 194.0000 194.0535 196.0000 196.0000 196.0000
#> [11,] 128.9567 128.000 128.0867 130.0867 132.0000 131.9567 130.7833 125.9133
#> [12,] 164.8300 162.000 162.0000 162.2833 167.2267 171.2833 176.4533 183.9150
#> [13,] 202.8600 200.790 197.5800 191.5800 185.6500 180.6500 175.7209 171.7324
#> [14,]       NA      NA       NA       NA       NA       NA       NA       NA
#>         [,205]   [,206]   [,207]   [,208]   [,209]   [,210]    [,211]    [,212]
#>  [1,] 137.4800 128.8100 121.6512 120.5459 120.0451 124.0950 128.59500 133.09500
#>  [2,]       NA       NA       NA       NA       NA       NA        NA        NA
#>  [3,]       NA       NA       NA       NA       NA       NA        NA        NA
#>  [4,] 143.0500 143.9500 146.8500 149.8500 154.7500 154.0500 152.10000 149.15000
#>  [5,]       NA       NA       NA       NA 166.5929 159.9188 153.24472 146.25667
#>  [6,] 158.1400 157.0233 165.7900 179.6733 194.6500 208.6733 222.67333 233.74333
#>  [7,]       NA       NA       NA       NA       NA       NA        NA        NA
#>  [8,] 181.0133 185.0000 185.0133 189.0067 191.0183 196.5183 201.99667 200.98667
#>  [9,] 180.8750 173.3750 165.9667 164.0000 164.0000 164.0833 169.11667 176.13333
#> [10,] 196.0600 198.0600 200.0900 203.0900 206.0299 206.9734 206.04698 208.00000
#> [11,] 124.0000 123.8700 120.7400 114.7400 108.8267 104.6533  96.65449  88.83946
#> [12,] 182.4150 180.4900 171.3200 159.6600 153.9435 152.8395 150.00000 150.28333
#> [13,] 167.7900 164.7900 161.2300 150.1600 138.5800 132.7900 130.00000 130.21000
#> [14,]       NA       NA       NA       NA       NA       NA        NA        NA
#>          [,213]   [,214]    [,215]    [,216]    [,217]   [,218]   [,219]
#>  [1,] 137.59500 145.7350 154.23500 156.82000 158.82000 157.8625 156.6125
#>  [2,]        NA       NA        NA        NA        NA       NA       NA
#>  [3,]        NA       NA        NA        NA        NA       NA       NA
#>  [4,] 145.20000 146.4250 147.92500 148.95000 151.85000 153.9000 160.6500
#>  [5,] 142.14667 136.2200 130.22000 127.11000 121.70167 116.2017       NA
#>  [6,] 249.62667 259.7667 261.95333 265.90667 266.97667 264.0700 258.1400
#>  [7,]        NA       NA        NA        NA        NA       NA       NA
#>  [8,] 196.99000 193.9900 191.00000 191.00667 192.99667 192.0033 192.9933
#>  [9,] 184.18333 195.1500 204.03333 206.15000 215.03333 217.0833 222.0167
#> [10,] 208.00000 208.0000 207.97000 207.00000 206.94000 205.0000 204.9700
#> [11,]  84.82667  81.0000  81.21667  86.04333  87.21667  92.2600  98.2600
#> [12,] 155.34000 161.1700 164.17000 167.05667 168.05667 168.8300 165.7167
#> [13,] 132.79000 129.7900 126.72000 122.86000 121.35000 126.0000 126.4200
#> [14,]        NA       NA        NA        NA        NA       NA       NA
#>         [,220]   [,221]   [,222]   [,223]   [,224]   [,225]   [,226]   [,227]
#>  [1,] 155.3625 154.1125 152.1800 149.2609 150.8267 146.4485 140.5217 138.1733
#>  [2,]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [3,]       NA       NA       NA       NA       NA 111.0607 113.2607 115.4607
#>  [4,] 163.8500 161.1500 154.3500 145.4500 140.2658 137.1505 133.2000 132.0500
#>  [5,]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [6,] 250.1867 245.1167 232.7917 220.2917 206.3267 196.2333 186.2333 185.0233
#>  [7,]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [8,] 190.9833 185.9667 175.9600 163.9701 155.0000 158.0000 161.0000 164.0000
#>  [9,] 223.0667 226.9500 223.9333 219.9667 217.9167 212.9667 211.0333 213.0667
#> [10,] 203.8800 199.9700 198.9700 197.7600 189.6700 178.5200 162.6400 150.6100
#> [11,] 104.3033 111.2167 116.3467 124.3900 133.2600 139.3033 146.3900 155.1300
#> [12,] 160.8300 157.8300 154.9433 154.0567 155.0000 155.1700 158.2833 163.2833
#> [13,] 132.7000 142.6300 151.7000 161.5600 169.2800 173.4200 179.2800 183.0000
#> [14,]       NA       NA       NA       NA       NA       NA       NA       NA
#>         [,228]   [,229]   [,230]   [,231]   [,232]   [,233]   [,234]   [,235]
#>  [1,] 136.1733 137.0960 138.2960 139.4960 140.6960 141.8960 149.7633 158.2633
#>  [2,]       NA       NA       NA       NA 104.1467 101.2200       NA       NA
#>  [3,] 117.6607 119.8607 119.0633 117.1204 115.7467 114.4133 113.0800 111.1200
#>  [4,] 133.9000 135.5851 137.2536 138.9221 139.0000 139.0000 139.9533 139.0498
#>  [5,]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [6,] 183.0467 179.0933 177.0467 179.9300 180.9767 187.8367 192.8997 194.9600
#>  [7,]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [8,]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [9,] 217.1000 223.0667 227.0166 227.9066 220.9037 213.9008 206.8979 199.8950
#> [10,] 137.7000 127.7600 119.7600 111.8800 107.9100 104.9400 102.9400 100.9400
#> [11,] 158.1733 162.0000 162.1733 166.3467 174.4333 184.2167 189.1733 193.0000
#> [12,] 168.0567 169.0567 170.0000 170.0567 170.8867 169.0000 169.1133 170.9433
#> [13,]       NA       NA       NA       NA       NA       NA       NA       NA
#> [14,]       NA       NA       NA       NA       NA       NA       NA       NA
#>         [,236]    [,237]    [,238]    [,239]    [,240]    [,241]    [,242]
#>  [1,] 162.6533 164.82667 167.12875 169.45949 171.79023 172.00000 172.91181
#>  [2,]       NA        NA        NA        NA        NA        NA        NA
#>  [3,] 109.1200 107.12000 109.82000 113.76000 118.68439 121.81000 122.00000
#>  [4,] 139.0000 140.90635 145.76667 147.90033 151.81271 155.81333 153.14000
#>  [5,]       NA        NA        NA        NA        NA        NA        NA
#>  [6,] 190.1163 187.06020 178.18000 167.22000 156.22000 142.28000 131.22000
#>  [7,]       NA        NA        NA        NA        NA        NA        NA
#>  [8,]       NA 222.98497 218.47746 213.99333 212.99333 211.94000 202.90000
#>  [9,] 192.8920 185.88912 178.88620 171.85539 163.17909 154.50278 145.91000
#> [10,]  99.0300  99.99333  99.77103  99.54872  99.32642  99.10411  98.88181
#> [11,] 193.0433 194.04333 195.17333 198.91333 197.08667 199.34783 207.00000
#> [12,] 170.2833 175.17000 178.00000 177.71667 172.49000 163.60333 156.83000
#> [13,]       NA        NA        NA        NA        NA        NA        NA
#> [14,]       NA        NA        NA        NA        NA        NA        NA
#>         [,243]   [,244]    [,245]    [,246]    [,247]    [,248]   [,249]
#>  [1,] 173.9101 174.0000 170.34667 161.78000 152.78000 144.69333 140.3467
#>  [2,]       NA       NA        NA        NA        NA        NA       NA
#>  [3,] 122.0000 120.1200 118.12000 115.18000 112.18000 109.19868 103.3624
#>  [4,] 153.0000 155.8595 157.90667 157.04667 154.14950 158.76589 159.0000
#>  [5,]       NA       NA        NA 108.10000 107.03333 109.32000 111.7200
#>  [6,] 121.2000 121.0000 125.90000 126.00000 121.10000 120.02000 116.2760
#>  [7,]       NA       NA        NA        NA        NA        NA       NA
#>  [8,] 187.9867 185.9600 179.96000 173.96000 167.99333 166.96013 160.9900
#>  [9,] 141.4100 136.9203 132.94983 129.94020 126.99165 126.49082 126.0399
#> [10,]  98.6595  98.4372  98.21489  97.96667  96.96667  95.86711  91.9097
#> [11,] 207.0467 207.9533 206.95333 205.81395 201.73913 195.67333 188.8133
#> [12,] 154.0569 155.1800 158.18000 161.00000 160.76000 156.82000 153.8200
#> [13,]       NA       NA        NA        NA        NA        NA       NA
#> [14,]       NA       NA        NA        NA        NA        NA       NA
#>          [,250]   [,251]    [,252]    [,253]   [,254]   [,255]   [,256]
#>  [1,] 137.71667 135.2167 131.34667 152.00667 152.1733 145.6067 145.9133
#>  [2,]        NA       NA        NA        NA       NA       NA       NA
#>  [3,] 100.65000  98.1500  96.43333  94.76667  93.1000  91.5900  90.0900
#>  [4,] 153.28000 147.2800 140.34884 137.14047 134.1495 135.9000 133.1405
#>  [5,] 114.12000 116.5200 118.92000 126.25000 133.7500 134.9667       NA
#>  [6,] 112.47600 108.6760 104.87600 101.07600 101.0000  99.0400 100.9600
#>  [7,]        NA       NA        NA 159.94000 158.5100 157.0100 157.0000
#>  [8,] 157.96000 151.9467 143.96000 137.97667 134.4767 130.9800 127.9733
#>  [9,] 128.00556 128.3393 128.67297 129.01000 129.5100 130.0399 132.0000
#> [10,]  88.96678  88.0301  89.13289  93.27090 102.2326 109.2100 116.0000
#> [11,] 184.81333 180.7200 174.58000 165.58000 156.5800 147.4400 135.8133
#> [12,] 150.94000 149.9400 148.82000 145.88000 144.0000 143.9400 143.0600
#> [13,]        NA       NA        NA 168.78930 166.2200 169.3667 174.3667
#> [14,]        NA       NA        NA        NA       NA       NA       NA
#>         [,257]   [,258]   [,259]   [,260]   [,261]   [,262]   [,263]   [,264]
#>  [1,] 145.5441 145.0449 139.5217 133.9850 128.4942 128.0000 127.5433 127.0433
#>  [2,]       NA       NA       NA 107.0000 107.0000 107.0000 107.0000 107.0000
#>  [3,]  88.1200  86.1200  85.0600  84.0600  83.0600  82.0600  81.0600  81.9400
#>  [4,] 126.3488 121.2341 116.2333 111.2333 108.1400 110.8600 114.8133 118.8133
#>  [5,]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [6,] 108.8400 117.8200 120.9400 118.0600 119.9600 124.9000 134.8000 135.0000
#>  [7,] 157.9933 155.0200 153.0133 150.0200 146.0267 142.0267 137.0333 132.0333
#>  [8,] 123.9733 119.9933 119.0000 118.9867 116.9933 115.9933 114.9667 109.9667
#>  [9,] 132.0083 132.5083 133.0542 136.3042 139.5542 142.8042 146.0833 151.0334
#> [10,] 116.0000 115.8796 111.9003 108.9399 106.9366 104.9333 102.9667 101.9667
#> [11,] 131.8133 128.0933 129.9067 127.9533 127.0000 126.9533 125.8133 121.7200
#> [12,] 144.0000 143.8800 141.9400 140.8800 138.8200 135.8800 133.8200 130.8200
#> [13,] 179.4767 185.9767 192.4385 197.7191 193.8533 191.7067 187.8533 185.8533
#> [14,]       NA       NA       NA       NA       NA       NA       NA       NA
#>          [,265]    [,266]    [,267]    [,268]    [,269]    [,270]    [,271]
#>  [1,] 128.82667 129.60889 130.27556 130.94222 133.74000 134.91333 131.34667
#>  [2,] 107.00000 107.00000 107.00000 105.76444 104.43111 103.09778 103.00000
#>  [3,]  80.59235  79.09484  79.23436  79.48457  79.73478  79.98499  80.37600
#>  [4,] 116.14000 117.90667 119.90667 118.09333 115.14000 110.23333 106.18667
#>  [5,]        NA        NA        NA        NA        NA        NA        NA
#>  [6,] 143.79070 153.79933 154.98000 157.94000 161.92000 175.72000 186.78000
#>  [7,] 128.02667 125.02000 123.01333 119.02667 115.02667 110.03333 107.02000
#>  [8,] 104.99333 103.98667 102.00000 102.00662 103.00000 103.99333 103.02667
#>  [9,] 153.02000 153.98000 152.98000 151.96000 149.96013 147.96656 145.96000
#> [10,] 100.96781  99.96892  99.00000  98.97000  97.99250  97.74250  97.49250
#> [11,] 115.72000 109.67333 102.81333  98.86000  95.81333  91.86000  88.95333
#> [12,] 127.82000 124.94000 123.82000 120.94000 119.94000 118.82000 115.82000
#> [13,] 183.70667 179.70667 176.07333 177.07333 178.22000 181.29333 185.03667
#> [14,]        NA        NA        NA        NA        NA        NA        NA
#>          [,272]    [,273]    [,274]    [,275]    [,276]    [,277]    [,278]
#>  [1,] 127.34667 123.34667 119.34667 120.36772 121.86522 121.49426 120.93850
#>  [2,] 103.00000 103.00000 103.00000 103.00000 103.00000 103.00000 103.00000
#>  [3,]  80.77600  81.17600  81.57600  81.97600  81.06000  81.00000  88.52000
#>  [4,] 105.04667 105.00000 105.95333 106.95333 108.90667 110.90667 111.00000
#>  [5,]        NA 141.93266 138.13333 135.10000 135.96667 136.96667 137.96667
#>  [6,] 201.70000 205.89404 208.93960 209.00000 208.02000 201.14000 199.04000
#>  [7,] 103.02667  99.02667  95.02667  92.02000  89.02000  88.00667  84.02667
#>  [8,] 107.00000 107.00000 107.00667 108.00667 108.99333 107.98667 105.98667
#>  [9,] 143.98000 143.02000 144.06000 147.04000 149.00000 148.90000 143.86000
#> [10,]  97.24250  96.94000  94.96990  93.96667  93.06667  95.00000  95.03333
#> [11,]  87.95333  86.95333  85.90667  83.95333  83.18667  87.04667  88.09333
#> [12,] 113.00000 113.24000 117.30000 122.36000 128.12000 129.88000 127.88000
#> [13,] 185.53667 186.00000 186.80132 197.00000 196.48667 189.41333 181.48667
#> [14,]        NA        NA        NA        NA        NA        NA        NA
#>          [,279]    [,280]    [,281]    [,282]    [,283]    [,284]    [,285]
#>  [1,] 120.38273 119.82697 119.27121 118.71545 118.15969 117.60393 117.04817
#>  [2,] 103.00000 102.53667 102.03667  96.44000  92.29333  91.07333  90.07333
#>  [3,]  89.00000  89.93980  90.94000  91.00000  91.94000  92.94000  92.06000
#>  [4,] 111.95333 114.86000 120.72000 127.67333 132.76667 137.76667 143.72000
#>  [5,] 141.86667 140.06667 132.26667 123.30000 115.26667 109.20000 100.30000
#>  [6,] 199.98000 200.00000 201.96000 203.96000 200.08000 193.14000 185.16000
#>  [7,]  83.00667  84.98667  88.97333  92.97333  95.98000  98.98000  96.02000
#>  [8,] 103.99333 102.98667 100.98000  97.98000  94.99333  93.99333  93.00000
#>  [9,] 136.98000 135.98000 134.94000 131.94000 128.96000 126.96000 124.86000
#> [10,]  96.06667  97.96667  96.96667  95.96667  94.96667  93.96667  92.96667
#> [11,]  90.12444  92.79111  95.45778  98.09302 100.10851 102.61269 105.14000
#> [12,] 125.28477 113.89262 111.64000 105.76000 101.76000  97.70000  92.88000
#> [13,] 174.85333 172.85333 170.78000 167.92667 166.78000 163.63333 158.34000
#> [14,]        NA        NA        NA        NA        NA        NA        NA
#>          [,286]    [,287]    [,288]
#>  [1,] 117.91333 116.17333 114.17333
#>  [2,]  90.00000  90.00000  91.85333
#>  [3,]  92.94000  92.06000  91.06000
#>  [4,] 145.90667 146.00000 144.09333
#>  [5,]  92.26667  82.33333  78.13333
#>  [6,] 174.22000 164.20000 157.14000
#>  [7,]  96.99333  97.99333  96.01333
#>  [8,]  92.99333  91.98000  88.99333
#>  [9,] 117.92000 113.90000 108.92000
#> [10,]  92.03333  93.00000  93.03333
#> [11,] 108.37333 116.18605 119.91304
#> [12,]  90.64000  84.82000  82.12000
#> [13,] 149.26667 139.56000 133.92667
#> [14,]        NA        NA        NA
#> 
#> $actual_dates
#>  [1] "2015-06-06" "2015-06-07" "2015-06-08" "2015-06-09" "2015-06-10"
#>  [6] "2015-06-11" "2015-06-12" "2015-06-13" "2015-06-14" "2015-06-15"
#> [11] "2015-06-16" "2015-06-17" "2015-06-18" "2015-06-19"
#> 
#> $dt0
#> [1] 5
#>