Calculate all metrics in iglu

The function all_metrics runs all of the iglu metrics, and returns the results with one column per metric.

all_metrics(data, dt0 = NULL, inter_gap = 45, tz = "", timelag = 15, lag = 1)

Arguments

data

DataFrame object with column names "id", "time", and "gl".

dt0

The time frequency for interpolation in minutes, the default will match the CGM meter's frequency (e.g. 5 min for Dexcom).

inter_gap

The maximum allowable gap (in minutes) for interpolation. The values will not be interpolated between the glucose measurements that are more than inter_gap minutes apart. The default value is 45 min.

tz

A character string specifying the time zone to be used. System-specific (see as.POSIXct), but " " is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning.

timelag

Integer indicating the time period (# minutes) over which rate of change is calculated. Default is 15, e.g. rate of change is the change in glucose over the past 15 minutes divided by 15.

lag

Integer indicating which lag (# days) to use. Default is 1.

Value

A tibble object with 1 row per subject and one column per metric is returned.

Details

All iglu functions are calculated within the all_metrics function, and the resulting tibble is returned with one row per subject and a column for each metric. Time dependent functions are calculated together using the function optimized_iglu_functions. For metric specific information, please see the corresponding function documentation.

Examples

data(example_data_1_subject)
all_metrics(example_data_1_subject)
#> Warning: Gap found in data for subject id: Subject 1, that
#>     exceeds long moving average. Double check glucose data.
#> # A tibble: 1 × 49
#>   id       ADRR  COGI    CV  eA1C   GMI GRADE GRADE_eugly GRADE_hyper GRADE_hypo
#>   <fct>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>       <dbl>       <dbl>      <dbl>
#> 1 Subjec…  15.2 0.930  26.9  5.94  6.27  3.47        28.2        71.7      0.221
#> # … with 39 more variables: HBGI <dbl>, LBGI <dbl>, hyper_index <dbl>,
#> #   hypo_index <dbl>, IGC <dbl>, IQR <dbl>, J_index <dbl>, M_value <dbl>,
#> #   MAD <dbl>, MAGE <dbl>, above_140 <dbl>, above_180 <dbl>, above_250 <dbl>,
#> #   below_54 <dbl>, below_70 <dbl>, in_range_63_140 <dbl>,
#> #   in_range_70_180 <dbl>, range <int>, SD <dbl>, Min. <dbl>, 1st Qu. <dbl>,
#> #   Median <dbl>, Mean <dbl>, 3rd Qu. <dbl>, Max. <dbl>, Conga <dbl>,
#> #   GVP <dbl>, MODD <dbl>, SD.Roc <dbl>, CV_Measures_Mean <dbl>, …

# Specify the meter frequency and change the interpolation gap to 30 min
all_metrics(example_data_1_subject, dt0 = 5, inter_gap = 30)
#> Warning: Gap found in data for subject id: Subject 1, that
#>     exceeds long moving average. Double check glucose data.
#> # A tibble: 1 × 49
#>   id       ADRR  COGI    CV  eA1C   GMI GRADE GRADE_eugly GRADE_hyper GRADE_hypo
#>   <fct>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>       <dbl>       <dbl>      <dbl>
#> 1 Subjec…  15.2 0.930  26.9  5.94  6.27  3.47        28.2        71.7      0.221
#> # … with 39 more variables: HBGI <dbl>, LBGI <dbl>, hyper_index <dbl>,
#> #   hypo_index <dbl>, IGC <dbl>, IQR <dbl>, J_index <dbl>, M_value <dbl>,
#> #   MAD <dbl>, MAGE <dbl>, above_140 <dbl>, above_180 <dbl>, above_250 <dbl>,
#> #   below_54 <dbl>, below_70 <dbl>, in_range_63_140 <dbl>,
#> #   in_range_70_180 <dbl>, range <int>, SD <dbl>, Min. <dbl>, 1st Qu. <dbl>,
#> #   Median <dbl>, Mean <dbl>, 3rd Qu. <dbl>, Max. <dbl>, Conga <dbl>,
#> #   GVP <dbl>, MODD <dbl>, SD.Roc <dbl>, CV_Measures_Mean <dbl>, …