Optimized Calculations of the iglu Functions: AUC, CONGA, GVP, MODD, SD_ROC, CV_MEASURES, & SD_MEASURES

The function all_metrics_optomized optimizes the functions AUC, CONGA, GVP, MODD, SD_ROC, CV_MEASURES, & SD_MEASURES iglu functions by extracting the CGMS2DayByDay calculation and passing the result into each function.

optimized_iglu_functions(data)

Arguments

data
DataFrame object with column names "id", "time", and "gl".

Value

If a data.frame object is passed, then a tibble object with 1 row for each subject, and fifteen columns is returned: a column for subject id, a column for CONGA(24) value, a column for Glucose Variability Percentage (GVP) value, a column for mean difference between glucose values obtained at the same time of day (MODD) value, a column for area under curve (AUC) value, a column for Coefficient of Variation mean (CV_Measures_Mean) value, a column for Coefficient of Variation standard deviation (CV_Measures_SD) value, a column for Mean Absolute Glucose (MAG) value, a column for Standard Deviation of rate of change (SD Roc) value, a column for Standard Deviation vertical within days (SdW) value, a column for Standard Deviation between time points (SdHHMM) value, a column for Standard Deviation within series (SdWSH) value, a column for horizontal Standard Deviation (SdDM) value, a column for Standard Deviation between days, within timepoints (SdB) value, a column for Standard Deviation between days, within timepoints value, corrected for changes in daily means (SdBDM) value.

Details

Returns a tibble object with 1 row for each subject, and fifteen columns: a column for subject id, a column for CONGA(24) value, a column for Glucose Variability Percentage (GVP) value, a column for mean difference between glucose values obtained at the same time of day (MODD) value, a column for area under curve (AUC) value, a column for Coefficient of Variation mean (CV_Measures_Mean) value, a column for Coefficient of Variation standard deviation (CV_Measures_SD) value, a column for Standard Deviation of rate of change (SD Roc) value, a column for Mean Absolute Glucose (MAG) value, a column for Standard Deviation vertical within days (SdW) value, a column for Standard Deviation between time points (SdHHMM) value, a column for Standard Deviation within series (SdWSH) value, a column for horizontal Standard Deviation (SdDM) value, a column for Standard Deviation between days, within timepoints (SdB) value, a column for Standard Deviation between days, within timepoints value, corrected for changes in daily means (SdBDM) value.

Examples

data(example_data_1_subject)
optimized_iglu_functions(example_data_1_subject)

#> # A tibble: 1 x 15
#>   id    Conga   GVP  MODD SD.Roc CV_Measures_Mean CV_Measures_SD   AUC   MAG
#> * <fct> <dbl> <dbl> <dbl>  <dbl>            <dbl>          <dbl> <dbl> <dbl>
#> 1 Subj…  37.0  1.17  27.8  0.620             21.1           7.80  123.  17.0
#> # … with 6 more variables: SdW <dbl>, SdHHMM <dbl>, SdWSH <dbl>, SdDM <dbl>,
#> #   SdB <dbl>, SdBDM <dbl>

data(example_data_5_subject)
optimized_iglu_functions(example_data_5_subject)

#> # A tibble: 5 x 15
#>   id    Conga   GVP  MODD SD.Roc CV_Measures_Mean CV_Measures_SD   AUC   MAG
#> * <fct> <dbl> <dbl> <dbl>  <dbl>            <dbl>          <dbl> <dbl> <dbl>
#> 1 Subj…  37.0  1.17  27.8  0.620             21.1           7.80  123.  17.0
#> 2 Subj…  60.6  1.21  44.1  0.642             17.0           6.41  218.  18.9
#> 3 Subj…  63.4  1.26  48.2  0.831             27.1           9.40  154.  27.1
#> 4 Subj…  33.5  1.17  24.9  0.617             18.4           5.70  130.  16.7
#> 5 Subj…  73.8  1.38  59.4  1.05              29.0           7.56  174.  35.3
#> # … with 6 more variables: SdW <dbl>, SdHHMM <dbl>, SdWSH <dbl>, SdDM <dbl>,
#> #   SdB <dbl>, SdBDM <dbl>