Plot meal metrics visualization

The function plot_meals produces a visual for meals data

Usage

plot_meals(data, mealtimes, plot_type=c('ggplot','plotly'))

Arguments

data

DataFrame object with column names "id", "time", and "gl". Should only be data for 1 subject. In case multiple subject ids are detected, a warning is produced and only 1st subject is used.

mealtimes

Either a vector of mealtimes, corresponding to data being from a single subject, OR a dataframe with at least 2 columns labeled id and mealtime. Optionally the mealtimes dataframe can include a column labeled meal, giving the meal type (helps to compensate for overlapping meals)

plot_type

One of "ggplot" or "plotly". Default: "gglplot". Returns either a ggplot (static image) or Plotly chart (interactive GUI).

Value

Plot to visualize meals data.

Details

Only a single subject's data may be plotted. The solid black line is the glucose trace. Vertical dashed red lines show the mealtimes, and the horizontal blue lines show the baseline for each meal. Purple triangles are plotted to illustrate the 3 meal_metrics Namely the three vertices show the baseline, peak, and 1hr post-peak recovery. If plot_type = 'plotly', plotly is used to display an interactive visual that allows one to zoom into specific areas of the plot.

References

Service, F. John. (2013) Glucose Variability, Diabetes 62(5): 1398-1404, doi:10.2337/db12-1396

See also

Author

Elizabeth Chun

Examples

data(example_data_hall)
data(example_meals_hall)
meal_metrics(example_data_hall, example_meals_hall)
#> No mealtimes found for subject: 1636-69-001
#> No mealtimes found for subject: 1636-69-026
#> No mealtimes found for subject: 1636-69-032
#> No mealtimes found for subject: 1636-69-090
#> No mealtimes found for subject: 1636-69-091
#> No mealtimes found for subject: 1636-69-114
#> No mealtimes found for subject: 1636-70-1005
#> No mealtimes found for subject: 1636-70-1010
#> No mealtimes found for subject: 2133-015
#> No mealtimes found for subject: 2133-017
#> No mealtimes found for subject: 2133-019
#> No mealtimes found for subject: 2133-021
#> No mealtimes found for subject: 2133-024
#> No mealtimes found for subject: 2133-027
#> No mealtimes found for subject: 2133-035
#> No mealtimes found for subject: 2133-036
#> Warning: There was 1 warning in `dplyr::reframe()`.
#> ℹ In argument: `meal_metrics_single(...)`.
#> ℹ In group 9: `id = "2133-004"`.
#> Caused by warning in `meal_metrics_single()`:
#> ! Some meals don't match with CGM readings. Mealtimes returned with NA for metric values
#> # A tibble: 25 × 11
#>    id           time                meal  deltag deltat basereco basegl peakgl
#>    <chr>        <dttm>              <chr>  <dbl>  <dbl>    <dbl>  <dbl>  <dbl>
#>  1 1636-69-001  NA                  NA      NA       NA   NA        NA     NA 
#>  2 1636-69-026  NA                  NA      NA       NA   NA        NA     NA 
#>  3 1636-69-032  NA                  NA      NA       NA   NA        NA     NA 
#>  4 1636-69-090  NA                  NA      NA       NA   NA        NA     NA 
#>  5 1636-69-091  NA                  NA      NA       NA   NA        NA     NA 
#>  6 1636-69-114  NA                  NA      NA       NA   NA        NA     NA 
#>  7 1636-70-1005 NA                  NA      NA       NA   NA        NA     NA 
#>  8 1636-70-1010 NA                  NA      NA       NA   NA        NA     NA 
#>  9 2133-004     2016-09-23 10:10:00 CF 1    87.8     80    0.362   135.   223.
#> 10 2133-004     2016-09-27 09:40:00 PB 1    NA       NA   NA        NA     NA 
#> # ℹ 15 more rows
#> # ℹ 3 more variables: recovergl <dbl>, peaktime <dttm>, recovertime <dttm>