Calculate percentage of time CGM was active

The function active_percent produces a tibble object with values equal to the percentage of time the cgm was active. For example, if a cgm's (5 min frequency) times were 0, 5, 10, 15 and glucose values were missing at time 5, then percentage of time the cgm was active is 75 The output columns correspond to the subject id and the precentage of time for which the cgm was active, and the output rows correspond to the subjects. The values will be between 0

active_percent(data, freqCGM = 5)

Arguments

data
DataFrame object with column names "id", "time", and "gl"

freqCGM
Numeric value of CGM Frequency(In minutes). Default value is 5.

Value

If a data.frame object is passed, then a tibble object with two columns: subject id and corresponding active_percent value is returned

Details

A tibble object with 1 row for each subject, a column for subject id and a column for active_percent values is returned.

References

Danne et al. (2017) International Consensus on Use of Continuous Glucose Monitoring Diabetes Care 40 .1631-1640, doi: 10.2337/dc17-1600 .

Examples

data(example_data_1_subject)

active_percent(example_data_1_subject)

#> # A tibble: 1 x 2
#>   id        active_percent
#>   <fct>              <dbl>
#> 1 Subject 1           79.8

active_percent(example_data_1_subject, freqCGM = 15)

#> # A tibble: 1 x 2
#>   id        active_percent
#>   <fct>              <dbl>
#> 1 Subject 1           239.

data(example_data_5_subject)

active_percent(example_data_5_subject)

#> # A tibble: 5 x 2
#>   id        active_percent
#>   <fct>              <dbl>
#> 1 Subject 1           79.8
#> 2 Subject 2           98.2
#> 3 Subject 3           92.1
#> 4 Subject 4           98.7
#> 5 Subject 5           95.8

active_percent(example_data_5_subject, freqCGM = 5)

#> # A tibble: 5 x 2
#>   id        active_percent
#>   <fct>              <dbl>
#> 1 Subject 1           79.8
#> 2 Subject 2           98.2
#> 3 Subject 3           92.1
#> 4 Subject 4           98.7
#> 5 Subject 5           95.8