Calculate percentage below targeted values

#' @description The function below_percent produces a tibble object with values equal to the percentage of glucose measurements below target values. The output columns correspond to the subject id followed by the target values and the output rows correspond to the subjects. The values will be between 0 (no measurements) and 100 (all measurements).

below_percent(data, targets_below = c(50, 80))

Arguments

data
DataFrame with column names ("id", "time", and "gl"), or numeric vector of glucose values.

targets_below
Numeric vector of glucose thresholds. Glucose values from data argument will be compared to each value in the targets_below vector. Default list is (50, 80).

Value

If a data.frame object is passed, then a tibble object with a column for subject id and then a column for each target value is returned. If a vector of glucose values is passed, then a tibble object without the subject id is returned. as.numeric() can be wrapped around the latter to output a numeric vector.

Details

A tibble object with 1 row for each subject, a column for subject id and column for each target value is returned. NA's will be omitted from the glucose values in calculation of percent.

References

Rodbard (2009) Interpretation of continuous glucose monitoring data: glycemic variability and quality of glycemic control, Diabetes Technology and Therapeutics 11 .55-67, doi: 10.1089/dia.2008.0132 .

Examples

data(example_data_1_subject)

below_percent(example_data_1_subject)

#> # A tibble: 1 x 3
#>   id        below_50 below_80
#>   <fct>        <dbl>    <dbl>
#> 1 Subject 1        0    0.652

below_percent(example_data_1_subject, targets_below = c(50, 100, 180))

#> # A tibble: 1 x 4
#>   id        below_100 below_180 below_50
#>   <fct>         <dbl>     <dbl>    <dbl>
#> 1 Subject 1      27.3      91.8        0

data(example_data_5_subject)

below_percent(example_data_5_subject)

#> # A tibble: 5 x 3
#>   id        below_50 below_80
#>   <fct>        <dbl>    <dbl>
#> 1 Subject 1   0         0.652
#> 2 Subject 2   0         0    
#> 3 Subject 3   0         0.913
#> 4 Subject 4   0.0273    2.05 
#> 5 Subject 5   0         1.13 

below_percent(example_data_5_subject, targets_below = c(80, 180))

#> # A tibble: 5 x 3
#>   id        below_180 below_80
#>   <fct>         <dbl>    <dbl>
#> 1 Subject 1      91.8    0.652
#> 2 Subject 2      26.4    0    
#> 3 Subject 3      81.7    0.913
#> 4 Subject 4      95.4    2.05 
#> 5 Subject 5      62.2    1.13