Calculates the number of Hypo/Hyperglycemic events as well as other statistics

The function episode calculation produces the number of Hypo/Hyperglycemic events as well as other statistics such as mean duration

episode_calculation(
  data,
  lv1_hypo = 100,
  lv2_hypo = 70,
  lv1_hyper = 120,
  lv2_hyper = 160,
  dur_length = 15
)

Arguments

data

DataFrame object with column names "id", "time", and "gl"

lv1_hypo

A double specifying a hypoglycemia threshold for level 1

lv2_hypo

A double specifying a hypoglycemia threshold for level 2

lv1_hyper

A double specifying a hyperglycemia threshold for level 1

lv2_hyper

A double specifying a hyperglycemia threshold for level 2

dur_length

An integer or a double specifying a duration length in minutes

Value

A dataframe with

Hypo_ep

A mean value that counts the number of hypogyclemia episodes per days

Hyper_ep

A mean value that counts the number of hypergyclemia episodes per days

hypo_duration

A mean value of the hypoglycemia durations per days

hyper_duration

A mean value of the hyperclycemia durations per days

low_alert

A mean percentage of time in level 1 and 2 hypoglycemic range

target_range

A mean percentage of time in target range

high_alert

A mean percentage of time in level 1 and 2 hyperglycemic ranges

hypo_min_avg

A mean percentage of time for the hypoglycemia per days

hyper_min_avg

A mean Percentage of time for the hyperglycemia per days

Author

Johnathan Shih, Jung Hoon Seo

Examples

episode_calculation(example_data_5_subject, lv1_hypo=100, lv1_hyper= 120)
#>                 id Hypo_ep Hyper_ep hypo_duration hyper_duration low_alert
#> 1  Subject 1 (Lv1)    2.79     2.64        111.26         196.90     23.91
#> 2  Subject 2 (Lv1)    0.08     1.15         50.00         964.62      0.27
#> 3  Subject 3 (Lv1)    1.14     3.71         66.67         287.91      7.36
#> 4  Subject 4 (Lv1)    2.43     4.71         93.29         218.66     15.27
#> 5  Subject 5 (Lv1)    1.00     2.42         96.56         534.49      8.74
#> 6  Subject 1 (Lv2)    0.07     1.50         25.00         111.62      0.16
#> 7  Subject 2 (Lv2)    0.00     2.08           NaN         600.29      0.00
#> 8  Subject 3 (Lv2)    0.14     2.71         25.00         154.13      0.25
#> 9  Subject 4 (Lv2)    0.14     2.57         25.00          87.95      0.39
#> 10 Subject 5 (Lv2)    0.08     4.00         15.00         172.26      0.09
#>    high_alert target_range hypo_min_avg hyper_min_avg
#> 1       43.90        32.18         0.24          0.44
#> 2       97.45         2.29         0.00          0.97
#> 3       81.16        11.47         0.07          0.81
#> 4       63.28        21.45         0.15          0.63
#> 5       82.41         8.85         0.09          0.82
#> 6       13.51        86.34         0.00          0.14
#> 7       85.09        14.91         0.00          0.85
#> 8       34.89        64.87         0.00          0.35
#> 9       17.55        82.07         0.00          0.18
#> 10      50.80        49.12         0.00          0.51