Calculates the number of Hypo/Hyperglycemic events as well as other statistics

The function episode calculation produces the number of Hypo/Hyperglycemic events as well as other statistics such as the average glucose level, mean duration, and the percentage of time spent in the ranges.

episode_calculation(data, hypo_thres=100.0, hyper_thres= 120.0, dur_length = 15)

Arguments

data
DataFrame object with column names "id", "time", and "gl"

hypo_thres
A double specifying a hypoglycemia threshold. The default value is set to hypo_thres = 100.0

hyper_thres
A double specifying a hyperglycemia threshold. The default value is set to hyper_thres = 120.0

dur_length
An integer or a double specfiying a duration length in minutes

Value

Data frame including Average Glucose, number of hypo and hyper episodes, Hypo and hyper mean values, the percentages of low, high and target alerts

Author

Johnathan Shih, Jung Hoon Seo

Examples

episode_calculation(example_data_5_subject, hypo_thres=100.0, hyper_thres= 120.0, dur_length = 15)

#> # A tibble: 5 x 11
#>   id     Average_Glucose Hypo_ep Hyper_ep hypo_duration hyper_duration Hypo_mean
#>   <fct>            <dbl>   <dbl>    <dbl>         <dbl>          <dbl>     <dbl>
#> 1 Subje…            125.  2.79       2.64         111.            197.    215.  
#> 2 Subje…            221.  0.0769     1.15          50             965.      7.32
#> 3 Subje…            156.  1.14       3.71          66.7           288.    106.  
#> 4 Subje…            132.  2.43       4.71          93.3           219.    206.  
#> 5 Subje…            172.  1          2.42          96.6           534.     91.1 
#> # … with 4 more variables: Hyper_mean <dbl>, low_alert <dbl>, high_alert <dbl>,
#> #   target_range <dbl>