Calculate Meal Metrics

The function meal_metrics calculates three simple glucose meal metrics

Usage

meal_metrics(data, mealtimes, before_win = 1, after_win = 3,
recovery_win = 1, interpolate = TRUE, adjust_mealtimes = TRUE, dt0 = NULL,
inter_gap = 45, tz = "")

Arguments

data

DataFrame object with column names "id", "time", and "gl". Should only be data for 1 subject. In case multiple subject ids are detected, the warning is produced and only 1st subject is used.

mealtimes

Either a vector of mealtimes, corresponding to data being from a single subject, OR a dataframe with at least 2 columns labeled id and mealtime. Optionally the mealtimes dataframe can include a column labeled meal, giving the meal type (helps to compensate for overlapping meals)

before_win

integer specifying number of hours to extend window before meal

after_win

integer specifying number of hours to extend window after meal

recovery_win

interger specifying number of hours for recovery beyond after window

interpolate

Boolean to indicate if CGM data should be interpolated or not. Default set to FALSE due to time intensive nature of interpolation. Parameters dt0, inter_gap, and tz will only be used if interpolate is set to TRUE.

adjust_mealtimes

Boolean to indicate if function should attempt to align mealtimes with CGM data times. This is important if mealtimes and CGM data times are not exactly aligned, because the function will return NA's for mealtimes that don't match with a corresponding CGM time stamp.

dt0

The time frequency for interpolation in minutes, the default will match the CGM meter's frequency (e.g. 5 min for Dexcom).

inter_gap

The maximum allowable gap (in minutes) for interpolation. The values will not be interpolated between the glucose measurements that are more than inter_gap minutes apart. The default value is 45 min.

tz

A character string specifying the time zone to be used. System-specific (see as.POSIXct), but " " is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning.

Value

A tibble object with 6 columns will be returned: id, time, meal, deltag, deltat, and basereco.

Details

A tibble object is returned with three metrics calculated for each mealtime. The last three columns of the output correspond to the three metrics: deltag refers to \(\Delta G\), deltat is \(\Delta T\), and basereco is % Baseline recovery. If no meal column is given in the original data, one will be automatically generated with a unique number for each meal.

References

Service, F. John. (2013) Glucose Variability, Diabetes 62(5): 1398-1404, doi:10.2337/db12-1396

Examples

data(example_data_hall)
data(example_meals_hall)
meal_metrics(example_data_hall, example_meals_hall)
#> # A tibble: 9 × 6
#>   id       time                meal  deltag deltat basereco
#>   <chr>    <dttm>              <chr>  <dbl>  <dbl>    <dbl>
#> 1 2133-010 2016-11-22 15:30:00 CF 1    42.1     60    0.741
#> 2 2133-010 2016-11-23 15:30:00 Bar 1   25.0     45    1.88 
#> 3 2133-010 2016-11-24 13:00:00 PB 1    37.2    155    0.918
#> 4 2133-013 2017-01-12 13:30:00 Bar 1   32.6     55    0.703
#> 5 2133-013 2017-01-16 15:00:00 CF 1    61.9     65    1.06 
#> 6 2133-013 2017-01-19 14:10:00 PB 1    10.2    120    0.109
#> 7 2133-018 2017-03-15 14:40:00 PB 1    97.8     80    0.378
#> 8 2133-018 2017-03-16 12:15:00 CF 1   171.      75    0.429
#> 9 2133-018 2017-03-17 14:05:00 Bar 1   77.6     75    0.670
meal_metrics(example_data_hall, example_meals_hall, interpolate = FALSE)
#> # A tibble: 9 × 6
#>   id       time                meal  deltag deltat basereco
#>   <chr>    <dttm>              <chr>  <dbl>  <dbl>    <dbl>
#> 1 2133-010 2016-11-22 15:28:42 CF 1    42.9   60    NA     
#> 2 2133-010 2016-11-23 15:28:38 Bar 1   25.8   45.0   1.90  
#> 3 2133-010 2016-11-24 12:58:34 PB 1    38.2  155.    0.917 
#> 4 2133-013 2017-01-12 13:30:08 Bar 1   32.7   55.0   0.704 
#> 5 2133-013 2017-01-16 14:59:45 CF 1    62.2   65    NA     
#> 6 2133-013 2017-01-19 14:09:28 PB 1    10.5  120.    0.0948
#> 7 2133-018 2017-03-15 14:40:00 PB 1    97.8   80     0.378 
#> 8 2133-018 2017-03-16 12:14:57 CF 1   171.    75    NA     
#> 9 2133-018 2017-03-17 14:04:52 Bar 1   78.4   75     0.664